addon/requirements.txt
pyserial
websocket
